+++ apps/api.planx.uk/modules/gis/controller.ts
@@ -4,6 +4,7 @@ import { ServerError } from "../../errors/serverError.js";
 import { findLpa } from "./service/findLpa/index.js";
 import type { FindLpaRequest } from "./service/findLpa/types.ts";
 import type { GetPlanningConstraintsSchemaRequest } from "./service/getPlanningConstraintsSchema/types.js";
+import { getLocalAuthorityArticleFourSchema } from "./service/helpers.js";
 
 export const findLpaController: FindLpaRequest = async (_req, res, next) => {
   const { lon, lat } = res.locals.parsedReq.query;
@@ -21,20 +22,24 @@ export const findLpaController: FindLpaRequest = async (_req, res, next) => {
   }
 };
 
-export const getPlanningConstraintsSchemaController: GetPlanningConstraintsSchemaRequest = async (_req, res, next) => {
-  const { localAuthority } = res.locals.parsedReq.query;
+export const getPlanningConstraintsSchemaController: GetPlanningConstraintsSchemaRequest =
+  async (_req, res, next) => {
+    const { localAuthority } = res.locals.parsedReq.query;
 
-  try {
-    const baseSchema = getValidSchemaValues("PlanningDesignation");
-    // TODO merge in granular a4s for applicable localAuthorities
-    
-    return res.status(200).json(baseSchema);
-  } catch (error) {
-    next(
-      new ServerError({
-        status: error instanceof ServerError ? error.status : undefined,
-        message: `Failed to get planning constraints schema (${localAuthority}). ${(error as Error).message}`,
-      }),
-    );
-  }
-};
+    try {
+      const baseSchema = getValidSchemaValues("PlanningDesignation");
+      const articleFourSchema =
+        getLocalAuthorityArticleFourSchema(localAuthority);
+
+      return res
+        .status(200)
+        .json([...(baseSchema ?? []), ...(articleFourSchema ?? [])].sort());
+    } catch (error) {
+      next(
+        new ServerError({
+          status: error instanceof ServerError ? error.status : undefined,
+          message: `Failed to get planning constraints schema (${localAuthority}). ${(error as Error).message}`,
+        }),
+      );
+    }
+  };
