+++ apps/api.planx.uk/modules/gis/docs.yaml
@@ -27,11 +27,11 @@ paths:
                 items:
                   type: object
   /planning-constraints-schema:
-    get: 
+    get:
       tags: ["gis"]
       summary: Get PlanX-supported data values for planning constraints
       description: Returns PlanX-supported data values for planning constraints in-service automations, including granular Article 4 values for applicable local authorities
-      parameters: 
+      parameters:
         - in: query
           name: localAuthority
           required: false
@@ -39,7 +39,7 @@ paths:
       responses:
         "200":
           description: A JSON array of data values
-          content: 
+          content:
             application/json:
               schema:
                 type: array
